requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
groq>=0.11.0
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
pytest>=7.4.3
pytest-cov>=4.1.0
sentence-transformers>=2.2.2
pydantic>=2.0
langchain-groq>=0.1.0